refactored, added some input structure check and "hexlified" return values.
