+++ packages/bundler/src/UserOpMethodHandler.ts
@@ -2,15 +2,19 @@ import { BigNumber, ethers, Wallet } from 'ethers'
 import { JsonRpcProvider, JsonRpcSigner, Provider } from '@ethersproject/providers'
 import { BundlerConfig } from './BundlerConfig'
 import { EntryPoint } from './types'
-import { UserOperationStruct } from './types/contracts/BundlerHelper'
 import { hexValue, resolveProperties } from 'ethers/lib/utils'
 import { rethrowError } from '@account-abstraction/utils'
 import { calcPreVerificationGas } from '@account-abstraction/sdk/dist/src/calcPreVerificationGas'
 import { debug_traceCall } from './GethTracer'
 import { BundlerCollectorReturn, bundlerCollectorTracer } from './BundlerCollectorTracer'
 import Debug from 'debug'
+import { UserOperationStruct } from '@account-abstraction/contracts'
+import { UserOperationEventEvent } from '@account-abstraction/contracts/dist/types/EntryPoint'
+import { deepHexlify, requireCond } from './utils'
 
 const debug = Debug('aa.handler.userop')
+
+const HEX_REGEX = /^0x[a-fA-F\d]/i
 
 export class UserOpMethodHandler {
   constructor (
@@ -47,6 +51,24 @@ export class UserOpMethodHandler {
     return beneficiary
   }
 
+  async validateUserOperation (userOp1: UserOperationStruct, requireSignature = true) {
+
+    //minimal sanity check: userOp exists, and all members are hex
+    requireCond(userOp1 != null, 'No UserOperation param')
+    const userOp = await resolveProperties(userOp1) as any
+
+    const fieldNames = 'sender,nonce,initCode,callData,callGasLimit,verificationGasLimit,preVerificationGas,maxFeePerGas,maxPriorityFeePerGas,paymasterAndData'
+    let fields = fieldNames.split(',')
+    if (requireSignature) {
+      fields.push('signature')
+    }
+    fields.forEach(key => {
+      requireCond(userOp[key] != null, 'Missing userOp field: ' + key, -32602)
+      let value = userOp[key].toString()
+      requireCond(value.match(HEX_REGEX) != null, `Invalid ${key}:${value} in UserOp`, -32602)
+    })
+  }
+
   /**
    * simulate UserOperation.
    * Note that simulation requires debug API:
@@ -55,6 +77,9 @@ export class UserOpMethodHandler {
    * @param entryPointInput
    */
   async simulateUserOp (userOp1: UserOperationStruct, entryPointInput: string) {
+
+    await this.validateUserOperation(userOp1, false)
+    requireCond(entryPointInput != null, 'No entryPoint param')
 
     const userOp = await resolveProperties(userOp1)
     if (entryPointInput.toLowerCase() !== this.config.entryPoint.toLowerCase()) {
@@ -74,30 +99,27 @@ export class UserOpMethodHandler {
         gasLimit: simulationGas
       }, { tracer: bundlerCollectorTracer })
 
-      function require (cond: boolean, msg: string) {
-        if (!cond) throw new Error(msg)
-      }
-
       //todo: validate keccak, access
       //todo: block access to no-code addresses (might need update to tracer)
 
-      const bannedOpCodes = new Set([`GASPRICE`, `GASLIMIT`, `DIFFICULTY`, `TIMESTAMP`, `BASEFEE`, `BLOCKHASH`, `NUMBER`, `SELFBALANCE`, `BALANCE`, `ORIGIN`, `GAS`, `CREATE`, `COINBASE`])
+      const bannedOpCodes = new Set(['GASPRICE', 'GASLIMIT', 'DIFFICULTY', 'TIMESTAMP', 'BASEFEE', 'BLOCKHASH', 'NUMBER', 'SELFBALANCE', 'BALANCE', 'ORIGIN', 'GAS', 'CREATE', 'COINBASE'])
 
+      const paymaster = (userOp.paymasterAndData?.length ?? 0) >= 42 ? userOp.paymasterAndData.toString().slice(0, 42) : undefined
       const validateOpcodes = result.numberLevels['0'].opcodes
       const validatePaymasterOpcodes = result.numberLevels['1'].opcodes
       // console.log('debug=', result.debug.join('\n- '))
       Object.keys(validateOpcodes).forEach(opcode =>
-        require(!bannedOpCodes.has(opcode), `wallet uses banned opcode: '${opcode}`)
+        requireCond(!bannedOpCodes.has(opcode), `wallet uses banned opcode: ${opcode}`, 32501)
       )
       Object.keys(validatePaymasterOpcodes).forEach(opcode =>
-        require(!bannedOpCodes.has(opcode), `paymaster uses banned opcode: '${opcode}`)
+        requireCond(!bannedOpCodes.has(opcode), `paymaster uses banned opcode: ${opcode}`, 32501, { paymaster })
       )
       if (userOp.initCode.length > 2) {
-        require((validateOpcodes['CREATE2'] ?? 0) <= 1, 'initCode with too many CREATE2')
+        requireCond((validateOpcodes['CREATE2'] ?? 0) <= 1, 'initCode with too many CREATE2', 32501)
       } else {
-        require((validateOpcodes['CREATE2'] ?? 0) < 1, 'banned opcode: CREATE2')
+        requireCond((validateOpcodes['CREATE2'] ?? 0) < 1, 'banned opcode: CREATE2', 32501)
       }
-      require((validatePaymasterOpcodes['CREATE2'] ?? 0) < 1, 'paymaster uses banned opcode: CREATE2')
+      requireCond((validatePaymasterOpcodes['CREATE2'] ?? 0) < 1, 'paymaster uses banned opcode: CREATE2', 32501, { paymaster })
     }
   }
 
@@ -128,5 +150,33 @@ export class UserOpMethodHandler {
 
     // await postExecutionDump(this.entryPoint, requestId)
     return requestId
+  }
+
+  async _getUserOperationEvent (requestId: string): Promise<UserOperationEventEvent> {
+    const event = await this.entryPoint.queryFilter(this.entryPoint.filters.UserOperationEvent(requestId))
+    return event[0]
+  }
+
+  async getUserOperationReceipt (userOpHash: string) {
+    requireCond(userOpHash?.toString()?.match(HEX_REGEX) != null, 'Missing/invalid userOpHash', -32601)
+    const event = await this._getUserOperationEvent(userOpHash)
+    if (event == null) {
+      return null
+    }
+    const receipt = await event.getTransactionReceipt() as any
+    receipt.status = event.args.success ? 1 : 0
+    receipt.userOpHash = userOpHash
+    return deepHexlify(receipt)
+  }
+
+  async getUserOperationTransactionByHash (userOpHash: string) {
+    requireCond(userOpHash?.toString()?.match(HEX_REGEX) != null, 'Missing/invalid userOpHash', -32601)
+    const event = await this._getUserOperationEvent(userOpHash)
+    if (event == null) {
+      return null
+    }
+    const tx = await event.getTransaction() as any
+    tx.userOpHash = userOpHash
+    return deepHexlify(tx)
   }
 }
